- make sure to exit fullscreen mode (by releasing the d3d device) if our render initialisation code fails. If this isn't done, the error message box wont show as it'll be hidden behind the fullscreen window. - added 'lastError' string to the logging system so we can retrieve the last error message at any time - only add tab to log strings that we write to the file - put single quotes around some file paths to make error messages a bit clearer
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@484

+++ 3dc/avp/win95/Projload.cpp
@@ -2640,7 +2640,7 @@ texID_t load_rif_bitmap (char const * fname, BMPN_Flags flags)
 
 void release_rif_bitmaps()
 {
-	for (uint32_t i = 0; i < ingameTextureList.size(); i++)
+	for (size_t i = 0; i < ingameTextureList.size(); i++)
 	{
 		Tex_Release(ingameTextureList[i]);
 	}
